backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
supabase==2.10.0
anthropic==0.39.0
pydantic-settings==2.6.1
python-multipart==0.0.12
